refactor utils
- move utils to utils/ directory
- clean code on CMakeLists
