+++ src/CMakeLists.txt
@@ -4,23 +4,27 @@ add_subdirectory (io)
 add_subdirectory (quantization)
 
 file (GLOB CPP_SRCS "*.cpp")
-file (GLOB CPP_FACTORY_SRCS "factory/*.cpp")
-file (GLOB CPP_CONJUGATE_GRAPH_SRCS "impl/*.cpp")
-file (GLOB CPP_INDEX_SRCS "index/*.cpp")
-file (GLOB CPP_HNSWLIB_SRCS "algorithm/hnswlib/*.cpp")
-file (GLOB CPP_DATA_CELL_SRCS "data_cell/*.cpp")
-file (GLOB CPP_ALGORITHM_SRCS "algorithm/*.cpp")
 list (FILTER CPP_SRCS EXCLUDE REGEX "_test.cpp")
-list (FILTER CPP_FACTORY_SRCS EXCLUDE REGEX "_test.cpp")
+
+file (GLOB CPP_CONJUGATE_GRAPH_SRCS "impl/*.cpp")
 list (FILTER CPP_CONJUGATE_GRAPH_SRCS EXCLUDE REGEX "_test.cpp")
+
+file (GLOB CPP_INDEX_SRCS "index/*.cpp")
 list (FILTER CPP_INDEX_SRCS EXCLUDE REGEX "_test.cpp")
+
+file (GLOB CPP_HNSWLIB_SRCS "algorithm/hnswlib/*.cpp")
+list (FILTER CPP_HNSWLIB_SRCS EXCLUDE REGEX "_test.cpp")
+
+file (GLOB CPP_DATA_CELL_SRCS "data_cell/*.cpp")
 list (FILTER CPP_DATA_CELL_SRCS EXCLUDE REGEX "_test.cpp")
+
+file (GLOB CPP_ALGORITHM_SRCS "algorithm/*.cpp")
 list (FILTER CPP_ALGORITHM_SRCS EXCLUDE REGEX "_test.cpp")
 
 file (GLOB CPP_UTILS_SRCS "utils/*.cpp")
 list (FILTER CPP_UTILS_SRCS EXCLUDE REGEX "_test.cpp")
 
-set (VSAG_SRCS ${CPP_SRCS} ${CPP_FACTORY_SRCS} ${CPP_INDEX_SRCS} ${CPP_CONJUGATE_GRAPH_SRCS} ${CPP_HNSWLIB_SRCS}
+set (VSAG_SRCS ${CPP_SRCS} ${CPP_INDEX_SRCS} ${CPP_CONJUGATE_GRAPH_SRCS}
         ${CPP_HNSWLIB_SRCS} ${CPP_DATA_CELL_SRCS} ${CPP_ALGORITHM_SRCS} ${CPP_UTILS_SRCS})
 
 add_library (vsag SHARED ${VSAG_SRCS})
