+++ src/index/brute_force_parameter.cpp
@@ -17,8 +17,8 @@
 
 #include <fmt/format-inl.h>
 
-#include "../utils.h"
 #include "inner_string_params.h"
+#include "utils/util_functions.h"
 #include "vsag/constants.h"
 namespace vsag {
