+++ src/index/hnsw.cpp
@@ -30,9 +30,10 @@
 #include "empty_index_binary_set.h"
 #include "index/hnsw_zparameters.h"
 #include "io/memory_block_io_parameter.h"
-#include "io/memory_io_parameter.h"
 #include "quantization/fp32_quantizer_parameter.h"
 #include "safe_allocator.h"
+#include "utils/slow_task_timer.h"
+#include "utils/timer.h"
 #include "vsag/binaryset.h"
 #include "vsag/constants.h"
 #include "vsag/errors.h"
