+++ src/parameter_check.cpp
@@ -18,7 +18,7 @@
 #include "index/diskann_zparameters.h"
 #include "index/hnsw_zparameters.h"
 #include "resource_owner_wrapper.h"
-#include "utils.h"
+#include "utils/util_functions.h"
 #include "vsag/constants.h"
 #include "vsag/errors.h"
 #include "vsag/expected.hpp"
