+++ src/utils/slow_task_timer.h
@@ -1,0 +1,33 @@
+
+// Copyright 2024-present the vsag project
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include <chrono>
+#include <string>
+
+namespace vsag {
+class SlowTaskTimer {
+public:
+    explicit SlowTaskTimer(std::string name, int64_t log_threshold_ms = 0);
+    ~SlowTaskTimer();
+
+public:
+    std::string name;
+    int64_t threshold;
+    std::chrono::steady_clock::time_point start;
+};
+
+}  // namespace vsag
