+++ src/utils/util_functions.h
@@ -15,65 +15,13 @@
 
 #pragma once
 
-#include <chrono>
-#include <cstdint>
 #include <string>
-#include <unordered_set>
-#include <vector>
 
-#include "allocator_wrapper.h"
 #include "index/index_common_param.h"
-#include "logger.h"
-#include "spdlog/spdlog.h"
 #include "vsag/errors.h"
 #include "vsag/expected.hpp"
 
 namespace vsag {
-
-struct SlowTaskTimer {
-    explicit SlowTaskTimer(std::string name, int64_t log_threshold_ms = 0);
-    ~SlowTaskTimer();
-
-    std::string name;
-    int64_t threshold;
-    std::chrono::steady_clock::time_point start;
-};
-
-struct Timer {
-    explicit Timer(double& ref);
-    ~Timer();
-
-    double& ref_;
-    std::chrono::steady_clock::time_point start;
-};
-
-class WindowResultQueue {
-public:
-    WindowResultQueue();
-
-    void
-    Push(float value);
-
-    [[nodiscard]] float
-    GetAvgResult() const;
-
-private:
-    size_t count_ = 0;
-    std::vector<float> queue_;
-};
-
-template <typename T>
-struct Number {
-    explicit Number(T n) : num(n) {
-    }
-
-    bool
-    in_range(T lower, T upper) {
-        return ((unsigned)(num - lower) <= (upper - lower));
-    }
-
-    T num;
-};
 
 template <typename IndexOpParameters>
 tl::expected<IndexOpParameters, Error>
@@ -97,27 +45,5 @@ try_parse_parameters(JsonType& param_obj, IndexCommonParam index_common_param) {
 
 std::string
 format_map(const std::string& str, const std::unordered_map<std::string, std::string>& mappings);
-
-class LinearCongruentialGenerator {
-public:
-    LinearCongruentialGenerator() {
-        auto now = std::chrono::steady_clock::now();
-        auto timestamp =
-            std::chrono::duration_cast<std::chrono::milliseconds>(now.time_since_epoch()).count();
-        current_ = static_cast<unsigned int>(timestamp);
-    }
-
-    float
-    NextFloat() {
-        current_ = (A * current_ + C) % M;
-        return static_cast<float>(current_) / static_cast<float>(M);
-    }
-
-private:
-    unsigned int current_;
-    static const uint32_t A = 1664525;
-    static const uint32_t C = 1013904223;
-    static const uint32_t M = 4294967295;  // 2^32 - 1
-};
 
 }  // namespace vsag
